Add files via upload
The domultisnr module will stop if it finds a high quality single channel sequence, thereby greatly speeding up the algorithm in the case where noise is low or modest.  If a high quality prior sequence exists, its ending peak times/rr interval will be used to start the single channel search in new module combsearch.  The "getcandpks" module can select the signal polarity that maximizes quality; this should be run for a few segments for a given channel and then the polarity permanently set because flipping polarity back and forth can cause problems with peak times because the offsets are different depending on polarity.